+++ src/web/Session.php
@@ -83,10 +83,6 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
      */
     protected static $_originalSessionModule = null;
     /**
-     * Polyfill for ini directive session.use-strict-mode for PHP < 5.5.2.
-     */
-    private static $_useStrictModePolyfill = false;
-    /**
      * @var string the name of the session variable that stores the flash message data.
      */
     public $flashParam = '__flash';
@@ -393,12 +389,10 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
      * The cookie parameters passed to this method will be merged with the result
      * of `session_get_cookie_params()`.
      * @param array $value cookie parameters, valid keys include: `lifetime`, `path`, `domain`, `secure` and `httponly`.
-     * Starting with Yii 2.0.21 `sameSite` is also supported. It requires PHP version 7.3.0 or higher.
-     * For security, an exception will be thrown if `sameSite` is set while using an unsupported version of PHP.
-     * To use this feature across different PHP versions check the version first. E.g.
+     * Starting with Yii 2.0.21 `sameSite` is also supported.
      * ```
      * [
-     *     'sameSite' => PHP_VERSION_ID >= 70300 ? yii\web\Cookie::SAME_SITE_LAX : null,
+     *     'sameSite' => yii\web\Cookie::SAME_SITE_LAX,
      * ]
      * ```
      * See https://owasp.org/www-community/SameSite for more information about `sameSite`.
@@ -421,14 +415,7 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
     {
         $data = $this->getCookieParams();
         if (isset($data['lifetime'], $data['path'], $data['domain'], $data['secure'], $data['httponly'])) {
-            if (PHP_VERSION_ID >= 70300) {
-                session_set_cookie_params($data);
-            } else {
-                if (!empty($data['samesite'])) {
-                    $data['path'] .= '; samesite=' . $data['samesite'];
-                }
-                session_set_cookie_params($data['lifetime'], $data['path'], $data['domain'], $data['secure'], $data['httponly']);
-            }
+            session_set_cookie_params($data);
         } else {
             throw new InvalidArgumentException('Please make sure cookieParams contains these elements: lifetime, path, domain, secure and httponly.');
         }
@@ -566,17 +553,9 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
      */
     public function setUseStrictMode($value)
     {
-        if (PHP_VERSION_ID < 50502) {
-            if ($this->getUseCustomStorage() || !$value) {
-                self::$_useStrictModePolyfill = $value;
-            } else {
-                throw new InvalidConfigException('Enabling `useStrictMode` on PHP < 5.5.2 is only supported with custom storage classes.');
-            }
-        } else {
-            $this->freeze();
-            ini_set('session.use_strict_mode', $value ? '1' : '0');
-            $this->unfreeze();
-        }
+        $this->freeze();
+        ini_set('session.use_strict_mode', $value ? '1' : '0');
+        $this->unfreeze();
     }
 
     /**
@@ -586,10 +565,6 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
      */
     public function getUseStrictMode()
     {
-        if (PHP_VERSION_ID < 50502) {
-            return self::$_useStrictModePolyfill;
-        }
-
         return (bool)ini_get('session.use_strict_mode');
     }