+++ tests/TestCase.php
@@ -203,21 +203,9 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
     protected function invokeMethod($object, $method, $args = [], $revoke = true)
     {
         $reflection = new ReflectionObject($object);
+
         $method = $reflection->getMethod($method);
-
-        // @link https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_reflectionsetaccessible
-        // @link https://wiki.php.net/rfc/make-reflection-setaccessible-no-op
-        if (PHP_VERSION_ID < 80100) {
-            $method->setAccessible(true);
-        }
-
         $result = $method->invokeArgs($object, $args);
-
-        // @link https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_reflectionsetaccessible
-        // @link https://wiki.php.net/rfc/make-reflection-setaccessible-no-op
-        if ($revoke && PHP_VERSION_ID < 80100) {
-            $method->setAccessible(false);
-        }
 
         return $result;
     }
@@ -236,21 +224,9 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
         while (!$class->hasProperty($propertyName)) {
             $class = $class->getParentClass();
         }
+
         $property = $class->getProperty($propertyName);
-
-        // @link https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_reflectionsetaccessible
-        // @link https://wiki.php.net/rfc/make-reflection-setaccessible-no-op
-        if (PHP_VERSION_ID < 80100) {
-            $property->setAccessible(true);
-        }
-
         $property->setValue($object, $value);
-
-        // @link https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_reflectionsetaccessible
-        // @link https://wiki.php.net/rfc/make-reflection-setaccessible-no-op
-        if ($revoke && PHP_VERSION_ID < 80100) {
-            $property->setAccessible(false);
-        }
     }
 
     /**
@@ -263,24 +239,13 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
     protected function getInaccessibleProperty($object, $propertyName, $revoke = true)
     {
         $class = new ReflectionClass($object);
+
         while (!$class->hasProperty($propertyName)) {
             $class = $class->getParentClass();
         }
+
         $property = $class->getProperty($propertyName);
-
-        // @link https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_reflectionsetaccessible
-        // @link https://wiki.php.net/rfc/make-reflection-setaccessible-no-op
-        if (PHP_VERSION_ID < 80100) {
-            $property->setAccessible(true);
-        }
-
         $result = $property->getValue($object);
-
-        // @link https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_reflectionsetaccessible
-        // @link https://wiki.php.net/rfc/make-reflection-setaccessible-no-op
-        if ($revoke && PHP_VERSION_ID < 80100) {
-            $property->setAccessible(false);
-        }
 
         return $result;
     }