+++ tests/framework/behaviors/AttributeTypecastBehaviorTest.php
@@ -88,10 +88,6 @@ class AttributeTypecastBehaviorTest extends TestCase
 
     public function testTypecastEnum(): void
     {
-        if (PHP_VERSION_ID < 80100) {
-            $this->markTestSkipped('Can not be tested on PHP < 8.1');
-        }
-
         $model = new ActiveRecordAttributeTypecastWithEnum();
 
         $model->status = StatusTypeString::Active;
@@ -106,10 +102,6 @@ class AttributeTypecastBehaviorTest extends TestCase
      */
     public function testTypecastEnumFromString(): void
     {
-        if (PHP_VERSION_ID < 80100) {
-            $this->markTestSkipped('Can not be tested on PHP < 8.1');
-        }
-
         $model = new ActiveRecordAttributeTypecastWithEnum();
         $model->status = 'active'; // Same as StatusTypeString::ACTIVE->value;
 
@@ -123,10 +115,6 @@ class AttributeTypecastBehaviorTest extends TestCase
      */
     public function testTypecastEnumFailWithInvalidValue(): void
     {
-        if (PHP_VERSION_ID < 80100) {
-            $this->markTestSkipped('Can not be tested on PHP < 8.1');
-        }
-
         $model = new ActiveRecordAttributeTypecastWithEnum();
         $model->status = 'invalid';