+++ tests/framework/db/mysql/connection/DeadLockTest.php
@@ -39,9 +39,6 @@ class DeadLockTest extends ConnectionTest
         if (\stripos($this->getConnection(false)->getServerVersion(), 'MariaDB') !== false) {
             $this->markTestSkipped('MariaDB does not support this test');
         }
-        if (PHP_VERSION_ID >= 70400 && PHP_VERSION_ID < 70500) {
-            $this->markTestSkipped('Stable failed in PHP 7.4');
-        }
         if (!\function_exists('pcntl_fork')) {
             $this->markTestSkipped('pcntl_fork() is not available');
         }
@@ -286,6 +283,7 @@ class DeadLockTest extends ConnectionTest
      */
     private function setErrorHandler(): void
     {
+        // TODO: PHP 8.2+ minimum — review and remove version skip
         if (PHP_VERSION_ID < 70000) {
             set_error_handler(function ($errno, $errstr, $errfile, $errline) {
                 throw new ErrorException($errstr, $errno, $errno, $errfile, $errline);