+++ tests/framework/helpers/FormatConverterTest.php
@@ -62,7 +62,7 @@ class FormatConverterTest extends TestCase
     {
         $this->assertEqualsAnyWhitespace('n/j/y, g:i A', FormatConverter::convertDateIcuToPhp('short', 'datetime', 'en-US'));
         $this->assertEquals(
-            PHP_VERSION_ID < 50600 ? 'd.m.y H:i' : 'd.m.y, H:i',
+            'd.m.y, H:i',
             FormatConverter::convertDateIcuToPhp('short', 'datetime', 'de-DE')
         );
     }
@@ -216,6 +216,7 @@ class FormatConverterTest extends TestCase
     public function testIntlIcuToJuiShortFormDateTime(): void
     {
         $this->assertEqualsAnyWhitespace('m/d/y, : ', FormatConverter::convertDateIcuToJui('short', 'datetime', 'en-US'));
+        // TODO: PHP 8.2+ minimum — review and remove version skip
         $this->assertEquals(
             PHP_VERSION_ID < 50600 ? 'dd.mm.y :' : 'dd.mm.y, :',
             FormatConverter::convertDateIcuToJui('short', 'datetime', 'de-DE')